added basic routing and each user page
