+++ src/API/fake.api/user.api.js
@@ -151,6 +151,14 @@ const fetchAll = () =>
     }, 1000)
   })
 
+const getById = id =>
+  new Promise(resolve => {
+    window.setTimeout(function () {
+      resolve(users.find(u => u._id === id))
+    }, 1000)
+  })
+
 export default {
-  fetchAll
+  fetchAll,
+  getById
 }
