+++ src/components/login.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+const Login = () => {
+  
+  return (
+    <>
+      <h2>Login</h2>
+    </>
+  )
+}
+
+Login.propTypes = {
+
+}
+
+export default Login
