+++ src/components/main.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+const Main = () => {
+
+  return (
+    <>
+      <h2>Main Page</h2>
+    </>
+  )
+}
+
+Main.propTypes = {
+
+}
+
+export default Main
