+++ src/components/navbar.jsx
@@ -1,0 +1,29 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+import { Link } from 'react-router-dom'
+
+const Navbar = () => {
+  
+  return (
+    <>
+      <ul className="nav">
+        <li className="nav-item">
+          <Link className="nav-link active" aria-current="page" to="/">Main</Link>
+        </li>
+        <li className="nav-item">
+          <Link className="nav-link" to="/login">Login</Link>
+        </li>
+        <li className="nav-item">
+          <Link className="nav-link" to="/users">Users</Link>
+        </li>
+      </ul>
+      
+    </>
+  )
+}
+
+Navbar.propTypes = {
+
+}
+
+export default Navbar
