+++ src/components/usersTable.jsx
@@ -3,10 +3,11 @@ import PropTypes from 'prop-types'
 import Bookmark from './bookmark'
 import QualitiesList from './qualitiesList'
 import Table from './table'
+import { Link } from 'react-router-dom'
 
 const UsersTable = ({ users, onSort, selectedSort, onMarking, onDelete }) => {
   const columns = {
-    name: { path: 'name', name: 'Имя' },
+    name: { path: 'name', name: 'Имя', component: user => <Link to={`/users/${user._id}`}>{user.name}</Link> },
     qualities: { name: 'Качества', component: user => <QualitiesList qualities={user.qualities} /> },
     professions: { path: 'profession.name', name: 'Профессия' },
     completedMeetings: { path: 'completedMeetings', name: 'Встретился, раз' },
